requirements.txt
onnx
onnxruntime
transformers
sentencepiece
torch
pillow
